Refactor lightning module into bionemo-llm (#123)
Refactors the `BioBertLightningModule` into a reusable, model kind-agnostic `BionemoLightningModule` abstract base class. This ABC is in `bionemo-llm`. Uses of this class should **parameterize** the construction of the class. They should **not** extend the class. 

`BioBertLightningModule` has been refactored to use this class. The class has been replaced by a function that constructs a `BioBertLightningModule` instance that uses the right tokenizer and model architecture. Its semantics are unchanged otherwise.