Make multi-line RUN statements fail fast (#225)
Makes each multi-line `RUN` in the Dockerfile fail immediately once any individual command exits with a nonzero exit code. Before, it was possible for a command to fail and the step would still be marked as successful so long as the _last_ command did not fail. In turn, this would make the image build look like it succeeded even when it should have been marked as a failure. Now, each multi-line `RUN` statement starts with `set -eo pipefail`.