Replace Launcher Script with Justfile & Standalone Scripts + Instructions for External (#239)
Full support for `just` is slated for the next release. While this PR adds the `justfile`, it also adds stand-alone scripts under `internal/scripts/` that external users can use for the development lifecycle. Instructions for these have been added to the README.md. Documentation for the `justfile` has been moved to `internal/README_justfile.md`.

--

Replaces the `launch.sh` script with a new `justfile` with commands that support the development lifecycle. New commands ('recipes' in `just`) allow for building the development or release image, checking the setup (checking installed programs & their versions, creating the `.env` file or reading it, etc.), running interactive programs on either image, or running `pytest` in the release image. These new recipes are:
```
    build-dev              # Builds the development image.
    build-release          # Builds the release image.
    run-dev cmd='bash'     # Runs an interactive program in the development bionemo image.
    run-release cmd='bash' # Runs an interactive program in the release bionemo image.
    setup                  # Checks for installed programs (docker, git, etc.), their versions, and grabs the latest cache image.
    test                   # Executes pytest in the release image.
```

The image building commands are executed as either `just build-dev` or `just build-release` while their corresponding container-running commands are `just run-{dev,release} (cmd)`, where `cmd` is the path to an executable program in the image (it defaults to `bash`, like `launch.sh` did before). The `just test` recipie runs `pytest` on all of the bionemo code & produces an HTML code coverage report, which is written to `htmlcov/` on the host machine (this directory is volume-mounted). This test command also runs the notebook tests. Note that `just setup` is a dependency of all of these commands: meaning the setup will _always_ be performed.

To support building a development image, the `Dockerfile` has been modified to have a new `development` target. This builds off of `dev`, but keeps the `3rdparty/` and `sub-packages/` code in `/workspace/bionemo2`. Additionally, it installs all of the local code (`bionemo-*` and `3rdparty/*`) as **editable** installs. This enables development work as the code can be mounted and modified _outside_ of the `dist-packages` folder, which is owned by root. 

Note that `just build-dev` builds this new `development` target while `just build-release` builds the `release` target, which is equivalent to building the entire image as `release` is the final stage.