bionemo version is set statically in one file, used throughout (#345)
Changes how the `version` attribute for all bionemo sub-packages is
assigned. Previously, the version string was computed from a the state
of `git tag` in the repository. Specifically, the tag for the current
commit -- or the last such tag -- would be used as the version. This
caused problems whenever the tag was not of the right format.
Additionally, it made it hard to pin down exactly where the version is
coming from.

A simpler mechanism is in place. Now, the version always comes from the
`VERSION` file at the root level. To make installation work as expected,
each `sub-package` has a symlink to the root-level `VERSION` file.

An additional change is made in this PR: all LICENSE files also now
symlink to the root-level LICENSE. This better ensures that all
sub-packages are using the same verbatim license text.

---------

Signed-off-by: Malcolm Greaves <mgreaves@nvidia.com>