Pbinder/lazy load scdl (#332)
I have added functionality for paginated loading of an anndata file.
This enables building up an in-memory sparse matrix structure of the
file when it is converted into the SingleCellMemMapDataset structure.
This used in the case where the anndata file will not fit into memory.

The user can specify a file size cutoff of paginated_load_cutoff to
specify the minimum anndata file size at which paginated loading will
occur. load_block_size refers to the number rows (cells) that will be
loaded into memory at a given time.

**A line by line analysis of memory usage with regular data loading vs.
the paginated data-loading with a block-size of 100_000 on an h5ad file
of 967 MB:**

[memory.txt](https://github.com/user-attachments/files/17514163/memory.txt)

 
Regular data-loading adds 4825.8 MB to memory. The majority of this
(3258.4 MB) is during the loading of the entire h5ad file into memory.
The paginated data-loading with a block size of 100_000 adds at most
1382 MB to memory. A block size of 10_000 adds at most 297.6 MB to
memory.

This is a line-by-line breakdown of regular loading vs. paginated
loading with a block size of 100_000.

---------

Signed-off-by: polinabinder1 <pbinder@nvidia.com>
Co-authored-by: Steven Kothen-Hill <148821680+skothenhill-nv@users.noreply.github.com>
Co-authored-by: Malcolm Greaves <malcolmgreaves@users.noreply.github.com>